Move dialogs out from VernWithSuggestions to NewEntry, since not used by RecentEntry.
