+++ src/components/DataEntry/DataEntryTable/VernWithSuggestions/VernWithSuggestions.tsx
@@ -7,31 +7,25 @@ import {
   withLocalize,
 } from "react-localize-redux";
 
-import { simpleWord, Word } from "../../../../types/word";
 import DupFinder from "../../../../goals/MergeDupGoal/DuplicateFinder/DuplicateFinder";
-import SenseDialog from "./SenseDialog";
-import VernDialog from "./VernDialog";
 
 interface VernWithSuggestionsProps {
   isNew?: boolean;
   isDisabled?: boolean;
   vernacular: string;
   vernInput?: React.RefObject<HTMLDivElement>;
-  updateVernField: (newValue: string) => Word[];
-  setActiveGloss: (newGloss: string) => void;
+  updateVernField: (newValue: string) => boolean;
   allVerns: string[];
   handleEnterAndTab: (e: React.KeyboardEvent) => void;
   updateWordId: (wordId?: string) => void;
   selectedWordId?: string;
   onBlur?: () => void;
+  openVernDialog?: () => void;
 }
 
 interface VernWithSuggestionsState {
   vernOpen: boolean;
-  senseOpen: boolean;
   suggestedVerns: string[];
-  dupVernWords: Word[];
-  selectedWord: Word;
 }
 
 /**
@@ -49,10 +43,7 @@ export class VernWithSuggestions extends React.Component<
     super(props);
     this.state = {
       vernOpen: false,
-      senseOpen: false,
       suggestedVerns: [],
-      dupVernWords: [],
-      selectedWord: { ...simpleWord("", ""), id: "" },
     };
   }
 
@@ -81,7 +72,12 @@ export class VernWithSuggestions extends React.Component<
     if (value) {
       suggestedVerns = [...this.autoCompleteCandidates(value)];
       if (suggestedVerns.length < this.maxSuggestions) {
-        const sortedVerns: string[] = [...this.props.allVerns].sort(
+        const viableVerns: string[] = this.props.allVerns.filter(
+          (vern: string) =>
+            this.suggestionFinder.getLevenshteinDistance(vern, value) <
+            this.suggestionFinder.maxScore
+        );
+        const sortedVerns: string[] = viableVerns.sort(
           (a: string, b: string) =>
             this.suggestionFinder.getLevenshteinDistance(a, value) -
             this.suggestionFinder.getLevenshteinDistance(b, value)
@@ -101,9 +97,9 @@ export class VernWithSuggestions extends React.Component<
   }
 
   handleSelection(value: string) {
-    const dupVernWords: Word[] = this.props.updateVernField(value);
-    if (dupVernWords.length > 0) {
-      this.setState({ vernOpen: true, dupVernWords });
+    const isDupVern: boolean = this.props.updateVernField(value);
+    if (isDupVern && this.props.openVernDialog) {
+      this.props.openVernDialog();
     } else {
       this.props.updateWordId();
     }
@@ -111,89 +107,46 @@ export class VernWithSuggestions extends React.Component<
 
   render() {
     return (
-      <React.Fragment>
-        <Autocomplete
-          freeSolo
-          disabled={this.props.isDisabled}
-          value={this.props.vernacular}
-          options={this.state.suggestedVerns}
-          onBlur={(_event: React.FocusEvent<HTMLDivElement>) => {
-            if (this.props.onBlur) {
-              this.props.onBlur();
+      <Autocomplete
+        freeSolo
+        disabled={this.props.isDisabled}
+        value={this.props.vernacular}
+        options={this.state.suggestedVerns}
+        onBlur={(_event: React.FocusEvent<HTMLDivElement>) => {
+          if (this.props.onBlur) {
+            this.props.onBlur();
+          }
+          if (this.props.selectedWordId === undefined)
+            this.handleSelection(this.props.vernacular);
+        }}
+        onChange={(_event, value) => {
+          // onChange is triggered when an option is selected
+          if (!value) value = "";
+          this.handleSelection(value);
+          this.updateSuggestedVerns(value);
+        }}
+        onInputChange={(_event, value) => {
+          // onInputChange is triggered by typing
+          this.props.updateVernField(value);
+          this.updateSuggestedVerns(value);
+          this.props.updateWordId();
+        }}
+        onKeyPress={(e: React.KeyboardEvent) => {
+          if (!this.state.vernOpen && (e.key === "Enter" || e.key === "Tab"))
+            this.props.handleEnterAndTab(e);
+        }}
+        renderInput={(params) => (
+          <TextField
+            {...params}
+            fullWidth
+            inputRef={this.props.vernInput}
+            label={
+              this.props.isNew ? <Translate id="addWords.vernacular" /> : ""
             }
-            if (this.props.selectedWordId === undefined)
-              this.handleSelection(this.props.vernacular);
-          }}
-          onChange={(_event, value) => {
-            // onChange is triggered when an option is selected
-            if (!value) value = "";
-            this.handleSelection(value);
-            this.updateSuggestedVerns(value);
-          }}
-          onInputChange={(_event, value) => {
-            // onInputChange is triggered by typing
-            const dupVernWords = this.props.updateVernField(value);
-            this.setState({ dupVernWords });
-            this.updateSuggestedVerns(value);
-            this.props.updateWordId();
-          }}
-          onKeyPress={(e: React.KeyboardEvent) => {
-            if (!this.state.vernOpen && (e.key === "Enter" || e.key === "Tab"))
-              this.props.handleEnterAndTab(e);
-          }}
-          renderInput={(params) => (
-            <TextField
-              {...params}
-              fullWidth
-              inputRef={this.props.vernInput}
-              label={
-                this.props.isNew ? <Translate id="addWords.vernacular" /> : ""
-              }
-              variant={this.props.isNew ? "outlined" : "standard"}
-            />
-          )}
-        />
-        {this.props.isNew && (
-          <VernDialog
-            open={this.state.vernOpen}
-            handleClose={(selectedWordId?: string) => {
-              this.setState({ vernOpen: false }, () => {
-                this.props.updateWordId(selectedWordId);
-              });
-              if (selectedWordId) {
-                let selectedWord: Word = this.state.dupVernWords.find(
-                  (word: Word) => word.id === selectedWordId
-                )!;
-                this.setState({
-                  selectedWord,
-                  senseOpen: true,
-                });
-              } else if (selectedWordId === "") {
-                let selectedWord: Word = {
-                  ...simpleWord(this.props.vernacular, ""),
-                  id: "",
-                };
-                this.setState({ selectedWord });
-              }
-            }}
-            vernacularWords={this.state.dupVernWords}
+            variant={this.props.isNew ? "outlined" : "standard"}
           />
         )}
-        {this.props.isNew && (
-          <SenseDialog
-            selectedWord={this.state.selectedWord}
-            open={this.state.senseOpen}
-            handleClose={(senseIndex: number) => {
-              if (senseIndex >= 0) {
-                this.props.setActiveGloss(
-                  this.state.selectedWord.senses[senseIndex].glosses[0].def
-                );
-              }
-              this.setState({ senseOpen: false });
-            }}
-          />
-        )}
-      </React.Fragment>
+      />
     );
   }
 }
